v4/vibe-ai-adoption/pyproject.toml
[project]
name = "vibe-ai-ops"
version = "0.1.0"
description = "5 AI operators powering Vibe's GTM — Temporal + LangGraph + Anthropic SDK"
requires-python = ">=3.12"
dependencies = [
    # Layer 1: Temporal
    "temporalio>=1.7.0",

    # Layer 2: LangGraph
    "langgraph>=0.3.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "langchain-anthropic>=0.3.0",
    "langsmith>=0.2.0",

    # Shared services
    "anthropic>=0.40.0",
    "slack-sdk>=3.27.0",
    "hubspot-api-client>=9.0.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",

    # Observability
    "langfuse>=2.0.0",
    "openlit>=1.0.0",

    # Database
    "psycopg[binary]>=3.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
